+++ src/MagonoteToolkit/Models/AISourceCodeReviewer.cs
@@ -1,16 +1,14 @@
 ﻿using OpenAI;
 using OpenAI.Chat;
-using OpenAI.Models;
 using System;
 using System.ClientModel;
-using System.Collections.ObjectModel;
 
 namespace MagonoteToolkit.Models
 {
     /// <summary>
     /// AIソースコードレビュークラス
     /// </summary>
-    internal class AISourceCodeReviewer
+    internal class AISourceCodeReviewer : AIControllerBase
     {
         //--------------------------------------------------
         // メソッド
@@ -59,36 +57,8 @@ namespace MagonoteToolkit.Models
             {
                 reviewResult = $"{Resources.Strings.Error}:{ex.Message}";
             }
+
             return reviewResult;
-        }
-
-        /// <summary>
-        /// モデルリスト取得処理
-        /// </summary>
-        /// <returns>モデルリスト</returns>
-        public static ObservableCollection<string> GetModels()
-        {
-            // OpenAI APIの設定
-            OpenAIClientOptions options = new() { Endpoint = new Uri(ApplicationSettings.ReadSettingsAIOpenAIAPIBaseUrl()) };
-            OpenAIClient client = new(new ApiKeyCredential(ApplicationSettings.ReadSettingsAIOpenAIAPIKey()), options);
-            OpenAIModelClient modelClient = client.GetOpenAIModelClient();
-
-            // モデルリストを取得
-            ObservableCollection<string> modelList = [];
-            try
-            {
-                OpenAIModelCollection models = modelClient.GetModels();
-                foreach (var model in models)
-                {
-                    modelList.Add(model.Id);
-                }
-            }
-            catch (Exception)
-            {
-                // エラーが発生した場合は空のリストを返す
-            }
-
-            return modelList;
         }
     }
 }
